requirements.txt
flask==2.0.1
requests==2.26.0
supabase==2.11.0
aiohttp==3.8.1
python-dotenv==0.19.2



